requirements.txt
genanki>=0.13.1
Pillow>=10.0.0
